Configurable prompt impression delay for faster tests; rename mocks.ts
